Fix: CUD 로직 제외
