+++ src/main/java/org/example/productservice/category/domain/Category.java
@@ -19,7 +19,7 @@ public class Category {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "category_id")
-	private Long id;
+	private Long categoryId;
 
 	@Column(nullable = false, length = 100)
 	private String categoryName;
