--- src/main/java/org/example/productservice/category/vo/in/AddCategoryRequestVo.java
@@ -1,27 +1,0 @@
-package org.example.productservice.category.vo.in;
-
-import org.example.productservice.category.dto.in.AddCategoryRequestDto;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class AddCategoryRequestVo {
-    private String categoryName;
-    private String parentCategoryUuid;
-
-    @Builder
-    public AddCategoryRequestVo(String categoryName, String parentCategoryUuid) {
-        this.categoryName = categoryName;
-        this.parentCategoryUuid = parentCategoryUuid;
-    }
-
-    public AddCategoryRequestDto toDto(AddCategoryRequestVo addCategoryRequestVo) {
-        return AddCategoryRequestDto.builder()
-                .categoryName(addCategoryRequestVo.getCategoryName())
-                .parentCategoryUuid(addCategoryRequestVo.getParentCategoryUuid())
-                .build();
-    }
-}
