--- src/main/java/org/example/productservice/category/vo/in/DeleteCategoryRequestVo.java
@@ -1,15 +1,0 @@
-package org.example.productservice.category.vo.in;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class DeleteCategoryRequestVo {
-
-	private String categoryUuid;
-
-	public DeleteCategoryRequestVo(String categoryUuid) {
-		this.categoryUuid = categoryUuid;
-	}
-}
