+++ src/main/java/org/example/productservice/llm/infrastructure/LLMVersionRepository.java
@@ -5,9 +5,7 @@ import java.util.List;
 import org.example.productservice.llm.domain.LLMVersion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdminLLMVersionRepository extends JpaRepository<LLMVersion, Long> {
-
-	boolean existsByLlmVersionName(String llmVersionName);
+public interface LLMVersionRepository extends JpaRepository<LLMVersion, Long> {
 
 	List<LLMVersion> findByLlmIdAndDeletedFalse(Long llmId);
 }
