+++ src/main/java/org/example/productservice/llm/presentation/LLMVersionController.java
@@ -2,17 +2,11 @@ package org.example.productservice.llm.presentation;
 
 import java.util.List;
 
-import org.example.productservice.llm.application.LLMVersionService;
-import org.example.productservice.llm.dto.in.AddLLMVersionRequestDto;
-import org.example.productservice.llm.vo.in.AddLLMVersionRequestVo;
-import org.example.productservice.llm.vo.out.GetLLMVersionListRequestVo;
 import org.example.productservice.common.response.BaseResponse;
-import org.springframework.web.bind.annotation.DeleteMapping;
+import org.example.productservice.llm.application.LLMVersionService;
+import org.example.productservice.llm.vo.out.GetLLMVersionListRequestVo;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -26,33 +20,8 @@ import lombok.RequiredArgsConstructor;
 @Tag(name = "LLM 버전 관리", description = "LLM 버전 관리 API")
 public class LLMVersionController {
 
-	private final LLMVersionService LLMVersionService;
+	private final LLMVersionService llmVersionService;
 
-	@Operation(summary = "LLM 버전 생성", description = "LLM 버전 생성")
-	@PostMapping
-	public BaseResponse<Void> addLLMVersion(
-		@RequestBody AddLLMVersionRequestVo addLLMVersionRequestVo) {
-
-		LLMVersionService.addLLMVersion(AddLLMVersionRequestDto.toDto(addLLMVersionRequestVo));
-		return new BaseResponse<>();
-	}
-
-	@Operation(summary = "LLM 버전 수정", description = "LLM 버전 수정")
-	@PutMapping("/{llmVersionId}")
-	public BaseResponse<Void> updateLLMVersion(
-		@PathVariable Long llmVersionId, @RequestBody AddLLMVersionRequestVo addLLMVersionRequestVo) {
-
-		LLMVersionService.updateLLMVersion(llmVersionId, AddLLMVersionRequestDto.toDto(addLLMVersionRequestVo));
-		return new BaseResponse<>();
-	}
-
-	@Operation(summary = "LLM 버전 삭제", description = "LLM 버전 삭제")
-	@DeleteMapping("/{llmVersionId}")
-	public BaseResponse<Void> deleteLLMVersion(@PathVariable Long llmVersionId) {
-
-		LLMVersionService.deleteLLMVersion(llmVersionId);
-		return new BaseResponse<>();
-	}
 
 	@Operation(summary = "LLM 버전 리스트 조회", description = """
 	LLM ID 매칭 정보
@@ -63,7 +32,7 @@ public class LLMVersionController {
 	public BaseResponse<List<GetLLMVersionListRequestVo>> getLLMVersionListByLLMId(@PathVariable Long llmId) {
 
 		return new BaseResponse<>(
-			LLMVersionService.getLLMVersionListByLLMId(llmId)
+			llmVersionService.getLLMVersionListByLLMId(llmId)
 				.stream()
 				.map(llmVersionDto -> GetLLMVersionListRequestVo.builder()
 					.llmVersionId(llmVersionDto.getLlmVersionId())
