--- src/main/java/org/example/productservice/llm/vo/in/AddLLMRequestVo.java
@@ -1,17 +1,0 @@
-package org.example.productservice.llm.vo.in;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class AddLLMRequestVo {
-
-	private String llmName;
-	private String llmType;
-
-}
