--- src/main/java/org/example/productservice/llm/vo/in/AddLLMVersionRequestVo.java
@@ -1,16 +1,0 @@
-package org.example.productservice.llm.vo.in;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Builder
-public class AddLLMVersionRequestVo {
-
-	private String llmVersionName;
-	private Long llmId;
-}
