--- src/main/java/org/example/productservice/llm/vo/in/DeleteLLMRequestVo.java
@@ -1,15 +1,0 @@
-package org.example.productservice.llm.vo.in;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@Getter
-@Builder
-@AllArgsConstructor
-public class DeleteLLMRequestVo {
-
-	private Long llmId;
-}
