--- src/main/java/org/example/productservice/llm/vo/in/UpdateLLMRequestVo.java
@@ -1,18 +1,0 @@
-package org.example.productservice.llm.vo.in;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class UpdateLLMRequestVo {
-
-	private Long llmId;
-	private String name;
-	private Long llmTypeId;
-	private String llmTypeName;
-}
